Test #160177531 Fetching of the User Profile
