+++ app/static/js/app.js
@@ -2,65 +2,6 @@
 
 
 
-/////////////////////////////////////////////////////
-//  FETCH USER PROFILE                             //
-/////////////////////////////////////////////////////
-
-document.addEventListener('DOMContentLoaded', () => {
-    const token = JSON.parse(localStorage.getItem('access_token'));
-    const ProfileUrl = '//api-thriller-diary.herokuapp.com/api/v1/users/profile';
-    fetch(`${ProfileUrl}`, {
-        method: "GET",
-        headers: {
-            "Content-Type": "application/json",
-            "access_token": token
-        }
-    })
-
-    .then((response)=>{
-        response.json().then((data) => {
-            const user = Object.values(data['Profile'])
-            console.log(user[0])
-            const userinfo = document.getElementById("name");
-            const usermail = document.getElementById("email");
-            userinfo.innerHTML = `${user[1]}!`;
-            usermail.innerHTML = `${user[2]}!`;
-
-        })})
-            .catch(err => console.log(err));
-
-
-});
-
-/////////////////////////////////
-// Number of entries user Has  //
-/////////////////////////////////
-
-document.addEventListener('DOMContentLoaded', () => {
-    const token = JSON.parse(localStorage.getItem('access_token'));
-    // get number of entries
-    const EntriesUrl = '//api-thriller-diary.herokuapp.com/api/v1/entries/';
-    fetch(`${EntriesUrl}`, {
-        method: "GET",
-        headers: {
-            "Content-Type": "application/json",
-            "access_token": token
-        }
-    })
-
-    .then((response)=>{
-        response.json().then((data) => {
-            let entries_number = document.getElementById('num_of_entries');
-            const entries = Object.values(data['Entries'])
-            if (entries.length == 0){
-                entries_number.innerHTML = 0;
-            }
-            else{
-                entries_number.innerHTML = entries.length;
-            }
-        })})
-            .catch(err => console.log(err));
-});
 ///////////////////////////////////
 //  USER LOGOUT                 //
 //////////////////////////////////
