+++ app/templates/profile.html
@@ -115,7 +115,10 @@
      </footer>
     
      <!-- js -->
+     <script src="{{ url_for('static', filename='js/number-of-entries.js') }}"></script>
      <script src="{{ url_for('static', filename='js/app.js') }}"></script>
+     <script src="{{ url_for('static', filename='js/fetch-profile.js') }}"></script>
+     
      <script src="{{ url_for('static', filename='js/jquery-3.2.1.min.js') }}"></script>
  </body>
  </html>
